Added buttons for shifting the section of time being displayed forward or back.
Currently shifts time range being looked at by a percentage of the
number of ticks currently being displayed (can also be thought of as the width
of the plot windows).

Also adds a box for specifying this percentage.

Left and right keyboard shortcuts may be used instead of buttons as well.

Merged from shiftButtons branch.