Changes to shift controls and status bar
When I saw that Perla added a status bar (good idea) I changed instances of setToolTip() that John had used when he used a toolbar instead of a menubar to setStatusTip(). But, the status messages didn't appear in the status bar.

I changed to using window.statusBar() to create the status, bar, and repositioned the shift controls at the end of the two sliders. Now the status messages appear.

I think the position of the shift controls is good, but ideally would prefer the < and > buttons to be half their current width. I'm still new to this stuff and don't know how to do that yet.

Also, from the moment I first saw MagPy4, I thought the font in the time controls next to the sliders was WAY TOO BIG. I reduced the font size from 14 to 11. Seriously, it was so big it was a distraction. I may try 10.