Add a checkbox for viewing edited data in dataDisplay win
Displays the most recently edited data values for a given vector,
and does not display it if it hasn't been edited.

If the edited data lists have different lengths (such as after
applying a filter to the main 3 vectors and a rotation to all
vectors), then the list of times with the widest range is chosen
and NaN is displayed for data vectors that don't have values
at a given time.

May still have to work out bugs with choosing the time range to
display if the one w/ widest range does not necessarily cover
the time values for all data vectors.

Also made a few tiny fixes for saving data as CSV files and
to the stats info in the top left corner.