Fixed issue w/ exporting SVG images from main window
It turns out the problem with exporting SVG files was that the tick
values are of a much larger magnitude than the y values and Qt's SVG
generator isn't the most precise. To fix this, I changed the tick data
being plotted to be relative to the smallest tick, e.g. the first tick
is now 0, second tick is (second tick - smallest tick), and so on. The
tick offset is stored in the window class and by axes/viewboxes.

However, this ONLY affects the underlying plotted data and NOT the actual
data being stored in the dictionaries or edits. Axis labels/strings are set
to be the same as the ones for the original data. Data doesn't appear to
ever be accessed directly from the plots and only from the dictionaries so
this seemed more efficient than recreating plots just to save an SVG file.

I tried to restrict the changes in how data is plotted to the plot objects
themselves and will work on compartmentalizing them further later on to
limit any bugs that could result from forgetting to use the tick offset in
the few places that use it outside of these objects.