Fixed issue w/ shifted plot elements in exported SVG files
The issue was in the way pyqtgraph determines the DPI value.
This has been fixed in later versions of pyqtgraph, but not
the version installed through pip. Unfortunately, this meant
copying over most of the code used for the exporter, at least
until the pip version is up to date.