Correct wave analysis calculations, except ellipticity
There were some issues with the frequency ranges, the range
of  fast fourier transformed data being looked at for the
computations, a few small differences from BX in dealing with
negative values, and with matrices not being transposed as
needed before applying transformations.

The values seem pretty close to BX's values for now, and these
differences might be from how MarsPy handles certain values
elsewhere in the program.

Still need to determine if elipm is different because of these
differences or if there is still a bug in those computations.