plotTrace no longer manually splits data into segments by errors
Instead of creating a new segment for each section of the plot
that is bounded by an error value, replacing all the errors in the
to-be-plotted data with NaN and using connect='finite' in the
PlotDataItem means that any segment (e.g. two consecutive points)
that is attached to a NaN value will not be plotted.

This means that each data segment with errors will be represented
by a single plotDataItem object even though the plot trace may
visually appear to be split into multiple segments wherever there
are error values.

This change was made because having too many segments makes the
program run very slow while it tries to add every item individually
and this way there are fewer segments created for data with a lot
of error values.

Data is still manually split according to time gaps, however.