Corrected a few bugs relating to plotting edited data
When I switched to plotting data with the time values shifted
by an offset to fix SVG exporting, it should've used the minimum
time from all data rather than the edited data's times b/c the
edited data might be start at a later time.

Similarly, for replacing errors in edited data with NaN when using
the plotTrace function, the new data and the original data that
contains the locations for all the errors might have different
sizes so I used the method for determining new times (in getTimes),
but on the actual data to get the sizes to match. I don't think
that is the most accurate method, however, for getting the new
times either, so that might be something to fix later on.