Modified how long the file name labels should be in plots
Sometimes the file names were so long that it affected how
plots fit within the window and cut part of them off.

Now the function for getting the filenames label uses an
estimate for how many characters can fit within the window's
width to determine when to use an ellipsis instead of using
the number of files that are open.