Added plot appearance modification pop-up window
Currently applies to only current data and pens are reset
if the plots are updated otherwise such as through plot menu
or update in Spectra.

Allows for modification of native plot titles/axis labels and line style/width

Changing width to > 1 tends to be slow but that is a Qt bug.