Modified underlying code for adjusting plots after resizing
Created classes to simply creation of colored labels next to
plots in the main window and for adding plots with their labels
into the main plot grid. I did this to reduce the complexity of
the code and to reduce the amount of hard-coded values the program
used when resizing and updating things like labels because this
created issues with overlapping text and with modifying these
objects elsewhere in the code.

The plot grid handles its own resize events now as well. However,
code for determining the font size for the main window's trace
labels remains the same as before because adjusting font size
to scale with window size in this case is a very complicated task
and the solution used before works well enough now w/o issues.