Added option for adding extra tick labels in main win
Currently accessible through a right-click menu; Adds
tick labels/strings matching the datastring's data value
at a given time tick.

Currently adds every set of additional labels to the top
of window above the main plots. Can easily be changed to
be at bottom but not sure if it would make things confusing.

This feature needs to be renamed to something better.