Added functionality for selecting multiple regions in main window
Users can now hold the Ctrl/Command key while clicking on the
plots in the main window to select more regions to get trace
stats info for.

Created new classes and restructured code for creating and
detecting the current state for selecting a section of the
plots.

Changes to the timeEdit in the traceStats window affect the region
that was last added, or the plot last moved, whichever event
was most recent. When the user moves a selected region, the
timeEdit values update accordingly to that region.

Behavior for right-clicks and left-clicks when regions have already
been added to main window (without ctrl key pressed) remain the
same. Having a single line open in the main window still gives
only value info as well and moving it still behaves like a single
InfiniteLine item until another point is selected to create a region.