Added option in Edit win for performing simple calculations on dta
Users can now enter a simple expression for modifying the original
data vectors using simple operands such as +,-,*,/,^ and add them
into the dataset like other edit operations.

Users must specify expressions in the form of 'VAR = EXPRESSION' where
VAR must be a datastring that has been loaded from the flat files
and EXPRESSION is a combination of the same (or other) datastring(s)
and/or numbers/operands that form a mathematical expression that can
be evaluated using the data vectors in place of datastrings.