Added tool to calculate normal vector using timing method in MagPy
Users select a time range appropriate for applying the timing method
and it will run the calculations as soon as this is selected.

Users can change the reference axis and threshold value, pressing
update to apply changes, similar to spectra tool.

Currently temporarily adds a line on all plots to indicate where
threshold value is to make it easier to use. Not sure if this is
needed.