Bug fixes to better support Galileo data and non-Insight/MMS files
Variables in axes selection boxes in the edit window and wave
analysis window now default to the first 3 variables in the
flat file. Also added fixed a bug where the edit window updated
the labels incorrectly when replotting data.

Internal restructuring done of the vector layout UI in the edit
window.