Internal changes to color bars used for spectrograms
Switched from using a fixed set of tick values that are drawn
over a graphics item to using a pyqtgraph AxisItem to take
advantage of the algorithms it uses to determine tick spacing
and values and better size adjustments when the window size
changes.

Switching to using a layout item to line up the axis item with
the color bar also makes it easier to set the offsets needed to
make the color bar line up with the plot's top/bottom edges.