Added color-map option to projection plots in trajectory analysis
Default 'projection onto the terminator plane' plots now map
each point to a color according to the time of the given sample.
Blue correlates to the start time and red correlates to the end
of the selected time.

A gradient legend with a DateTime axis is displayed in the
top-right corner of the plot.