Added option in Plot Menu to set a height stretch factor
Accessible under the small corner menu button for each sub-plot.

Stretch factor is relative to other factors for each sub-plot,
where the default stretch is 1. A stretch value of 2 or 3
means the sub-plot should have a height about 2 or 3 times
larger than the height of a sub-plot with a stretch value of 1.