Added option to export dynamic plot data
Dynamic Spectra, Dynamic Coh/Phase, and Dynamic Wave Analysis
plot data can be exported as TXT file with two grids in it.
Each grid has the frequencies as the first column and the start
time of each calculation as the column headers. One grid is of
the calculated values and the other is of the mapped colors in
hexadecimal format.

Accessible as a 'Save Plot Data' context menu option for the plots.