Initial implementation of saving/loading user workspace feature
Notes:

Files, edit history, general tools and selections are all reloaded.
Saved regions and batch select regions are also reloaded.

General tools such as Edit, Plot Menu, etc. do not currently
have their state saved. Also, costly operations such as calculating
the dynamic spectrogram grid values are all recalculated when
the workspace is loaded. These things may be improved in the future.

Saved regions and batch select regions are also reloaded.

Currently, the opened workspace is not automatically saved or
loaded as the default path when saving the workspace again, but
this may also be improved later.

Some error checking is performed in case the original files cannot
be opened or the workspace file itself cannot be opened.

Lastly, improvements need to be made to make sure tools are
always on top of the main window when they are reloaded.