Performance improvements for 'color' plots
Created a SpectraGridItem subclass to recreate the functionality
of stacked SpectraLine items, but optimized to significantly
improve the performance when loading many items.

Instead of plotting individual lines for each row in a spectrogram
plot, a single plot data item is generated instead that draws the
entire grid. This reduces the number of calls to expensive
functions related to painting and drawing paths and facilitates the
use of generators to process the subpaths more quickly.