Fixed issue w/ tick precision in MMS Formation tool
TimeEdits have times represented up to 3 decimals, but time ticks
from flat files may be more precise than that. This created issues
when checking if the interval was contained within the loaded
flat file's time range and resulted in downloading position data.
To work around this, the loaded data's time ticks are rounded to
3 decimal places for the purposes of comparing them to the user-set
time range.