Plots in main grid w/ many points are now downsampled by default
If > 2.5 million points are plotted in the main window, then
downsampling is enabled, which reduces the number of visible
samples. This does not affect the actual data stored. It improves
performance at the cost of reduced visual accuracy, which is
small if the dataset is very large. It can be turned off by
unchecking 'Downsample plot data' in the 'Options' menu.